requirements.txt
opencv-python
ultralytics
pyyaml
numpy
tk